Added mac address gathering and saving to the database, fixed some bugs in the traffic analyze service and controller

+++ src/server/controllers/trafficAnalyzeController.js
@@ -28,10 +28,23 @@ module.exports.getLogs = async function (request, _reply) {
 
 module.exports.removeCollection = async function (request, _reply) {
   const { rootCollectionId, docsCollectionId } = request.body
+  if (process.env.ACCEPT_REMOVE_COLLECTION === 'false') throw new HttpError[403]('Forbidden')
   const message = await trafficAnalyzesService.removeCollection(rootCollectionId, docsCollectionId)
 
   if (!message) {
     throw new HttpError[501]('Command execution failed')
+  }
+
+  return {
+    message
+  }
+}
+
+module.exports.macSaving = async function (_request, _reply) {
+  const message = await trafficAnalyzesService.macSaving()
+
+  if (!message) {
+    throw new HttpError[500]('Command execution failed')
   }
 
   return {
